day 3 part 1: refactor to reduce special cases
Instead of checking the rows and columns for not being first and
last, we can just add a boarder to our matrix. Also, to be
able to chain operations we can rework the part number
related functions to return an Option instead of bool. This
way we can use `and_then`.

Added some more helper functions so that we keep the same level
of abstraction in `add_part_numbers`.

Signed-off-by: Andreea Florescu <fandree@amazon.com>